[Jest] Migrate Attribution.js (#646)
* Move and rename

* Run jscodeshift

* rename

* manuell

* add before and after and change sourceDataType because of https://github.com/maplibre/maplibre-gl-js/commit/476ce360b6f27b0abae7a8db1c9dde6bb8629264#diff-9c473852318c57bac25989bc6b1e85a2dc3e53bceff7497b0314c407c6f9d036R114 ?

* eslint rules for snapshots

* add snapshots test for making sure, that the open attribut of the summary element ist handelt correct, see https://github.com/maplibre/maplibre-gl-js/issues/640
