+++ .eslintrc
@@ -103,6 +103,8 @@
     "jest/valid-expect": "error",
     "jest/prefer-to-be": "warn",
     "jest/no-alias-methods": "warn",
+    "jest/no-interpolation-in-snapshots": "warn",
+    "jest/no-large-snapshots": ["warn", { "maxSize": 50, "inlineMaxSize": 20 }],
     "jest/no-deprecated-functions": "warn"
   },
   "settings": {
